# docs/swagger.yaml
definitions:
  controller.listCouriers:
    properties:
      data:
        items:
          $ref: '#/definitions/dao.Courier'
        type: array
    type: object
...
  dao.Courier:
    properties:
      courier_name:
        type: string
      deleted:
        type: boolean
      delivery_service_id:
        type: integer
      email:
        type: string
      id_courier:
        type: integer
      number_of_failures:
        type: integer
      phone_number:
        type: string
      photo:
        type: string
      rating:
        type: integer
      ready_to_go:
        type: boolean
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
  dao.SmallInfo:
    properties:
      courier_name:
        type: string
      deleted:
        type: boolean
      id_courier:
        type: integer
      phone_number:
        type: string
      photo:
        type: string
      surname:
        type: string
    type: object
info:
  contact: {}
  description: Courier Service for Food Delivery Application
  title: Courier Service
paths:
  /courier:
    post:
      consumes:
      - application/json
      description: post new courier
      parameters:
      - description: Courier
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dao.Courier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dao.Courier'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: SaveCourier
      tags:
      - Courier
  /courier/{id}:
    get:
      consumes:
      - application/json
      description: get courier by ID
      parameters:
      - description: Courier ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dao.Courier'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: GetCourier
      tags:
      - Courier
    put:
      consumes:
      - application/json
      description: put courier status by courier ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: deleted
        in: body
        name: input
        required: true
        schema:
          type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dao.Courier'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: NewUpdateCourier
      tags:
      - Courier
  /couriers:
    get:
      consumes:
      - application/json
      description: get all couriers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dao.SmallInfo'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GetCouriers
      tags:
      - Couriers
  /couriers/photo:
    post:
      consumes:
      - image/jpeg
      description: set photo to DO Spaces, and it's way to DB
      parameters:
      - description: id courier
        in: query
        name: id
        required: true
        type: integer
      - description: logo image
        in: formData
        name: logo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: SaveCourierPhoto
      tags:
      - Couriers
  /couriers/service:
    get:
      description: get list of all couriers by courier service id
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: iddeliveryservice
        in: query
        name: iddeliveryservice
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.listCouriers'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: GetCouriersOfCourierService
      tags:
      - Couriers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"